Complete Odin Landing Page
Finished this project within two days, about 5 total hours
working on the project.

Not 100% one, since the page is missing key images, but the grand scheme
of things is done

This website will be used as my starting pad for The AC Doctors website.
Many things have been learned here. Read the README for more.